+++ src/CosmosDB/CosmosDB/Helpers/Constants.cs
@@ -326,5 +326,19 @@ namespace Microsoft.Azure.Commands.CosmosDB.Helpers
         public const string FleetspaceAccountObjectHelpMessage = "Fleetspace Account Object.";
         public const string GlobalDatabaseAccountResourceIdHelpMessage = "Resource ID of the global database account to associate with the fleetspace account.";
         public const string GlobalDatabaseAccountLocationHelpMessage = "Location of the global database account to associate with the fleetspace account.";
+
+        // Garnet Cluster constants
+        public const string GarnetClusterNameHelpMessage = "Name of the Garnet Cluster.";
+        public const string GarnetClusterObjectHelpMessage = "Garnet Cluster Object.";
+        public const string GarnetClusterLocationHelpMessage = "Azure Location of the Garnet Cluster.";
+        public const string GarnetClusterSubnetIdHelpMessage = "Resource id of a subnet that this cluster's management service should have its network interface attached to.";
+        public const string GarnetClusterReplicationFactorHelpMessage = "Number of copies of data maintained by the cluster.";
+        public const string GarnetClusterShardCountHelpMessage = "Number of shards in the cluster.";
+        public const string GarnetClusterNodeSkuHelpMessage = "Virtual Machine SKU used for clusters. Default value is Standard_DS14_v2.";
+        public const string GarnetClusterAvailabilityZoneHelpMessage = "If the data center has Availability Zone support, apply it to the Virtual Machine ScaleSet that hosts the garnet cluster virtual machines.";
+        public const string GarnetClusterAuthenticationMethodHelpMessage = "The authentication method used for the Garnet cluster. Acceptable values: 'Entra'.";
+        public const string GarnetClusterPersistenceHelpMessage = "Flag to indicate if persistence is enabled for the Garnet cluster.";
+        public const string GarnetClusterClusterTypeHelpMessage = "Type of the cluster. If set to Production, some operations might not be permitted on cluster. Acceptable values: 'Production', 'NonProduction'.";
+        public const string GarnetClusterExtensionsHelpMessage = "Extensions to be added or updated on the cluster.";
     }
 }